+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/NoSemiAfterFieldDeclarationRuleTest.java
@@ -1,0 +1,39 @@
+package org.sonar.plugins.delphi.pmd.rules;
+
+import static org.sonar.plugins.delphi.utils.conditions.RuleKeyAtLine.ruleKeyAtLine;
+
+import org.junit.Test;
+import org.sonar.plugins.delphi.utils.builders.DelphiTestUnitBuilder;
+
+public class NoSemiAfterFieldDeclarationRuleTest extends BasePmdRuleTest {
+
+  @Test
+  public void testFieldDeclarationsWithSemicolonsShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("type")
+            .appendDecl("  TType = class(TObject)")
+            .appendDecl("    Foo: TObject;")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
+  }
+
+  @Test
+  public void testFieldDeclarationsWithoutSemicolonsShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("type")
+            .appendDecl("  TType = class(TObject)")
+            .appendDecl("    Foo: TObject")
+            .appendDecl("  end;");
+
+    execute(builder);
+
+    assertIssues()
+        .areExactly(
+            1, ruleKeyAtLine("NoSemiAfterFieldDeclarationRule", builder.getOffsetDecl() + 3));
+  }
+}